lesson-3

+++ services/taskServices.js
@@ -1,44 +1,45 @@
 const HttpError = require("../utils/HttpError");
 const Task = require("../models/task");
 
-const getTasksService = async (page, limit, completed, owner) => {
+const getTasksService = async (page, limit, completed, userId) => {
   const skip = (page - 1) * limit;
-  // const filter = {};
-  // if (completed === "true") {
-  //   filter.completed = true;
-  // } else if (completed === "false") {
-  //   filter.completed = false;
-  // }
-  if (!completed) {
-    return await Task.find({ owner }).limit(limit).skip(skip);
+  const filter = {
+    owner: userId,
+  };
+  if (completed === "true") {
+    filter.completed = true;
+  } else if (completed === "false") {
+    filter.completed = false;
   }
-  return await Task.find({ owner, completed: completed })
-    .limit(limit)
-    .skip(skip);
+  return await Task.find(filter).skip(skip).limit(limit);
 };
 
-const getTaskService = async (taskId) => {
-  const task = await Task.findById(taskId);
+const getTaskService = async (taskId, userId) => {
+  const task = await Task.findOne({ _id, taskId, owner: userId });
   if (!task) {
     throw new HttpError(404, "Task not found");
   }
   return task;
 };
 
-const createTaskService = async (body, owner) => {
-  return await Task.create({ ...body, owner });
+const createTaskService = async (body, userId) => {
+  return await Task.create({ ...body, owner: userId });
 };
 
-const updateTaskService = async (taskId, data) => {
-  const updatedTask = await Task.findByIdAndUpdate(taskId, data, { new: true });
+const updateTaskService = async (taskId, data, userId) => {
+  const updatedTask = await Task.findByIdAndUpdate(
+    { _id: taskId, owner: userId },
+    data,
+    { new: true }
+  );
   if (!updatedTask) {
     throw new HttpError(404, "Task not found");
   }
   return updatedTask;
 };
 
-const deleteTaskService = async (taskId) => {
-  const task = await Task.findByIdAndDelete(taskId);
+const deleteTaskService = async (taskId, userId) => {
+  const task = await Task.findByIdAndDelete({ _id: taskId, owner: userId });
   if (!task) {
     throw new HttpError(404, "Task not found");
   }
